Move subsystems into separate package and fix references
