+++ src/main/java/frc/robot/subsystems/Drivetrain.java
@@ -1,21 +1,22 @@
-package frc.robot;
+package frc.robot.subsystems;
 
 import edu.wpi.first.wpilibj.command.Subsystem;
-import frc.robot.commands.TeleopDrive;
-import frc.robot.Utility;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.FeedbackDevice;
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.can.*;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
+import frc.robot.commands.TeleopDrive;
 import frc.robot.RobotMap;
 import frc.robot.Constants;
 import frc.robot.Robot;
+import frc.robot.DriveSignal;
+import frc.robot.Utility;
+import frc.robot.Util;
 
-
-public class Drivetrain extends Subsystem{
+public class Drivetrain extends Subsystem {
 	public double lastSpeed=0;
 	double _speedFactor = 1;
 	double _rotateFactor = 1;
